requirements.txt
aiogram==3.*
aiohttp
python-dotenv
translate
pydantic
matplotlib
numpy